Adapta codi per Django 4.2.1
El comportament del widget ClearableFileInput és diferent segons la
versió de Django.
